P11_Percobaan2_Pertanyaan1_23

+++ src/P11/Star23.java
@@ -7,7 +7,7 @@ public class Star23 {
         int n = 0;
         System.out.print("Masukkan nilai N : ");
         n = sc23.nextInt();
-        for (int i = 1; i <= n; i--) {
+        for (long i = 1; i <= n; i--) {
             System.out.print("*");
     }
     sc23.close();
